+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: Upgrade `libp2p` to `0.36.2` & `libp2p-gossipsub` to `0.13.0`. Some APIs are now async.
 - docs: Various improvements.
 - Ran `npm audit fix`.
+- `Waku.dial` accepts protocols expected from the peer. Defaults to Waku Relay only.
 
 ## [0.16.0] - 2022-01-31